fix: 일급컬렉션 적용 및 Map에서 모든 등수를 키값으로 포함하게 수정
